refactor: create connector service
